pyproject.toml
[tool.poetry]
name = "dset-3ftx"
version = "0.1.0"
description = ""
authors = ["tsenoner <tobias.senoner94@gmail.com>"]
readme = "README.md"
packages = [{include = "dset_3ftx"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.1"
pandas = "^1.5.3"
pyfaidx = "^0.7.1"
biopython = "^1.80"
requests = "^2.28.2"
tqdm = "^4.64.1"
openpyxl = "^3.1.0"
h5py = "^3.8.0"
distinctipy = "^1.2.2"
matplotlib = "^3.7.0"


[tool.poetry.group.dev.dependencies]
jupyter-black = "^0.3.3"
jupyter = "^1.0.0"
xlrd = "^2.0.1"
pycairo = "^1.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
